services/verifier/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.31.0
structlog==23.2.0
python-dotenv==1.0.0
